refactor: BoardViews 관련 로직 객체로 분리
